Move Raft command types to own file
server.go is getting long, so start breaking it up a little.
